% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userUtils.R
\name{chan_connect}
\alias{chan_connect}
\title{Estimate Channels Nodes Connect Through}
\usage{
chan_connect(i, j = NULL, model)
}
\arguments{
\item{i}{Node ids}

\item{j}{Node ids. If left blank, will select *all* edges with i}

\item{model}{LatClass model}
}
\description{
Estimates probability two nodes 
are connected through a channel, *conditional on them having an edge.*
}
\examples{
data("email_data")
mod = makeLatentModel(email_data$edgeList, 10, 
                      meta = email_data$meta)
mod$fit(fast_em = TRUE)

# Checking channel usage for 
# first few edges
nodes_1 = email_data$edgeList[1:5, 1]
nodes_2 = email_data$edgeList[1:5, 2]
chan_connect(nodes_1, nodes_2, mod)

# Checking channel usage for all edges 
# for first two nodes
chan_connect(i = c(1000, 1001), model = mod)
}
