% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userUtils.R
\name{channel_sizes}
\alias{channel_sizes}
\title{Compute sizes of channels}
\usage{
channel_sizes(mod, type = "nodes_using")
}
\arguments{
\item{mod}{LatClass model}

\item{type}{How size is defined. Either 'nodes_using' or 'exp_connects'}
}
\description{
Returns the size of each channel
}
\details{
The size of each channel can be defined in two ways: 
number of nodes that have non-zero attachment to a channel 
('nodes_using') *or* 
the expected number of connections through a channel a new node 
would have if it connected through that channel with probability 1
('exp_connects')
We note that the 'exp_connects' metric is a better description of size, 
but 'nodes_using' is more intuitive.
}
\examples{
data(email_data)
mod = makeLatentModel(email_data$edgeList, 10, 
                      metadata = email_data$meta)
mod$fit(fast_em = TRUE)

channel_sizes(mod, "exp_connects") 
}
