% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeTools.R
\name{checkMissingList}
\alias{checkMissingList}
\title{Check that there are no missing edges in original edge list}
\usage{
checkMissingList(obs_edges, missing_edges, max_n)
}
\description{
Check that there are no missing edges in original edge list
}
