% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userUtils.R
\name{computeExpConnects}
\alias{computeExpConnects}
\title{Compute expected connections for node}
\usage{
computeExpConnects(i, lcn_mod)
}
\arguments{
\item{i}{Index of node}

\item{lcn_mod}{LCN model}
}
\description{
Computes the estimated expected connections through each latent channel
}
