% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userUtils.R
\name{computeTheta}
\alias{computeTheta}
\title{Expected Channel Usage}
\usage{
computeTheta(i, j, lcn_mod)
}
\arguments{
\item{i}{Index of first node}

\item{j}{Index of second node}

\item{lcn_mod}{LCN model}
}
\description{
Compute expected channel usage between two connected nodes.
}
\details{
Computes the expected channel usage between two nodes 
\strong{conditional on the two nodes sharing an edge.}
}
