% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLCN.R
\name{emLCN}
\alias{emLCN}
\title{Optimization of Latent Channel Network via EM-style algorithms}
\usage{
emLCN(LCN_mod, iters = 10000, type = "EM", fast_em = T, a = 1,
  b = 1, tol = 10^-4, pTol = 10^-8)
}
\arguments{
\item{LCN_mod}{LCN model, output from makeLCN}

\item{iters}{Maximum iterations}

\item{type}{Algorithm type. Choices are "ECM", "EM" and "ParEM"}

\item{fast_em}{Logical: should fast EM update be used or basic?}

\item{tol}{Convergence tolerance}

\item{pTol}{Tolerance for skipping parameter updates}
}
\description{
Optimization of Latent Channel Network via EM-style algorithms
}
\details{
Fits a latent channel network with either an ECM algorithm 
(\code{type = "ECM"}), an EM algorithm (\code{type = "EM"}) or 
a parallel implementation of the EM algorithm (\code{type = "ParEM"}). 
In serial, the ECM algorithm tends to be fastest, 
but with at least two threads, the parallel EM algorithm 
is expected to be fastest. 

To control the number of threads used by the parallel EM algorithm, 
use \code{RcppParallel::setThreadOptions}.
}
