% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userUtils.R
\name{est_auc}
\alias{est_auc}
\title{Estimate Out-of-Sample AUC}
\usage{
est_auc(
  edgeList,
  models = c("LCN", "BKN"),
  nChan = 10,
  nEdgesMasked = 400,
  nNonEdgesMasked = 400
)
}
\arguments{
\item{edgeList}{nx2 matrix of edges}

\item{models}{Character vector of models to use}

\item{nChan}{Number of channels to use}

\item{nEdgesMasked}{Number of edges to mask}

\item{nNonEdgesMasked}{Number of non-edges to mask}
}
\description{
Estimate Out-of-Sample AUC
}
