% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userUtils.R
\name{get_both_auc}
\alias{get_both_auc}
\title{Get both in sample and out of sample AUC}
\usage{
get_both_auc(mod, out_edges, out_notEdges, in_edges, in_notEdges)
}
\description{
Get both in sample and out of sample AUC
}
