% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmapLCN}
\alias{heatmapLCN}
\title{Build heatmap from model}
\usage{
heatmapLCN(lcn_mod, grp, minGrpSize = NULL, cols = c("black",
  "lightblue", "orange", "red"), plotChannelNumber = T, xlab = " ",
  ylab = " ", sortColumns = T, ...)
}
\arguments{
\item{lcn_mod}{LCN model}

\item{grp}{Vector of group categories for each node}

\item{minGrpSize}{Minimum size of group in both. Smaller groups put in "other"}

\item{cols}{Colors for color gradient}

\item{xlab}{X-axis label}

\item{ylab}{Y-asix label}

\item{...}{Additional arguments passed to ComplexHeatmap::Heatmap}

\item{reorderChannels}{Should Channels be reorder by dependency on grp?}
}
\description{
Build heatmap from model
}
