% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBKN.R
\name{makeBKN}
\alias{makeBKN}
\title{Make Ball-Karrer-Newman (BKN) Model}
\usage{
makeBKN(edgeCountList, nDims = 5, unknownEdges = NULL)
}
\arguments{
\item{edgeCountList}{An nx3 matrix of edges. See details.}

\item{nDims}{Number of Latent Communities}

\item{unknownEdges}{An nx2 matrix of edges for which the count is unknown}
}
\description{
Make Ball-Karrer-Newman (BKN) Model
}
\details{
`edgeCountList` should be an nx3 matrix, with columns 1 & 2
being node ID's and column 3 being the number of edges. Note that 
the graph is considered to be undirected, so `[1,3,4]` and `[3,1,4]`
both imply that there are 4 edges between nodes 1 and 3. 
If node pairs appear more than once in edgeCountList, edge counts are summed.
}
