% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_net}
\alias{plot_net}
\title{Build heatmap from model}
\usage{
plot_net(
  mod,
  grp = NULL,
  metanames = NULL,
  minGrpSize = NULL,
  subset = NULL,
  name = "",
  ...
)
}
\arguments{
\item{mod}{LatClass object}

\item{grp}{Vector of group categories for each node}

\item{minGrpSize}{Minimum size of group in both. Smaller groups put in "other"}

\item{...}{Additional arguments passed to ComplexHeatmap::Heatmap}

\item{prob_cols}{Colors for color gradient of probability range}

\item{greater_col}{Color for color gradient beyond 1}

\item{reorderChannels}{Should Channels be reorder by dependency on grp?}

\item{xlab}{X-axis label}

\item{ylab}{Y-asix label}
}
\description{
Build heatmap from model
}
