% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatClass.R
\name{predict.LatClass}
\alias{predict.LatClass}
\title{Predictions from LatClass objects}
\usage{
\method{predict}{LatClass}(mod, i, j, type = "pairs")
}
\arguments{
\item{mod}{LatClass model}

\item{i}{node index}

\item{j}{Either an node index or metadata colname name}

\item{type}{Should node pairs or cross of all combinations be predicted}
}
\description{
Predict edge probabilities and categorical metadata
}
\examples{
data(email_data)
df = data.frame(dpt = email_data$nodeDpt)
# Grouping dpt for brevity
df$dpt[df$dpt > 4] = "other"

# Building model and fitting
mod = makeLatentModel(email_data$edgeList, 
                      nDims = 10, 
                      metadata = df)
mod$fit(fast_em = T)

# Predicting edge pairs
predict(mod, i = 1:3, j = 4:2)

# Predicting all combinations of i and j
predict(mod, i = 1:3, j = 1:3, type = "cross")

# Predicting metadata 
predict(mod, i = 1:3, "dpt")
}
