% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userUtils.R
\name{predicts_meta}
\alias{predicts_meta}
\title{Subsets channels that are predictive of metadata}
\usage{
predicts_meta(
  model,
  metanames = NULL,
  metavars = NULL,
  threshold = 0.5,
  sumFun = max
)
}
\arguments{
\item{model}{LatClass model}

\item{metanames}{Vector of names of metadata values to predict}

\item{metavars}{Vector of column names of metadata to predict}

\item{threshold}{Minimal parameter value to be considered predictive}

\item{sumFun}{Summary function: suggest either \code{max} or \code{sum}}
}
\description{
Returns both a vector of channels that are predictive of at least one
of the metadata values of interest and a parameter matrix of channels by 
metadata.
}
\details{
\code{metanames} refers to the individual values we might want to predict, 
while \code{metavars} is the column names.
}
\examples{
data(email_data)
mod = makeLatentModel(email_data$edgeList, 20,
                      meta = email_data$meta)
mod$fit(fast_em = TRUE)

# Returns channels that are predictive 
# of dpt == 1 or 2
predicts_meta(mod, metanames = c("dpt1", "dpt2") ) 
# Returns channels that are predictive 
# of *any* dpt
predicts_meta(mod, metanames = NULL, metavars = "dpt")
}
