% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBKN.R
\name{prepEdgeCountList}
\alias{prepEdgeCountList}
\title{Prep Edgelist}
\usage{
prepEdgeCountList(edgeCountList)
}
\arguments{
\item{edgeCountList}{An nx3 matrix of undirected edge pairs of the form 
n1, n2, count}
}
\description{
Prepares edgelist for \code{makeLCN}
}
