% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLCN.R
\name{prepEdgeList}
\alias{prepEdgeList}
\title{Prep Edgelist}
\usage{
prepEdgeList(edgeList, max_node = NULL)
}
\arguments{
\item{edgeList}{An nx2 matrix of undirected edge pairs}

\item{max_node}{Optional max node value: max(edgeList) used if null}
}
\description{
Prepares edgelist for \code{makeLCN}
}
