% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simLCN}
\alias{simLCN}
\title{Simulate Latent Channel Network}
\usage{
simLCN(p_mat)
}
\arguments{
\item{p_mat}{Matrix of channel usage probabilities}
}
\description{
Simulate Latent Channel Network
}
\examples{
data(email_data)

# Fitting model
mod = makeLatentModel(email_data$edgeList, 10)
mod$fit(fast_em = TRUE)
# Extracting fitted p-matrix
pmat_hat = mod$get_pars()$nodes
# Resampling according to given pmat_hat
new_edgeList = simLCN(pmat_hat)
}
