% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeTools.R
\name{unq_nondiag_flat}
\alias{unq_nondiag_flat}
\title{Return only unique non-selfloop samples from flat_index}
\usage{
unq_nondiag_flat(flat, max_n)
}
\description{
Return only unique non-selfloop samples from flat_index
}
