% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModelWrappers.R
\name{emLCN}
\alias{emLCN}
\title{Optimization of Latent Channel Network via EM}
\usage{
emLCN(LCN_mod, iters = 10000, tol = 10^-4, pTol = 10^-8)
}
\arguments{
\item{LCN_mod}{LCN model, output from makeLCN}

\item{iters}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{pTol}{Tolerance for skipping parameter updates}
}
\description{
Optimization of Latent Channel Network via EM
}
