% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmapLCN}
\alias{heatmapLCN}
\title{Build heatmap from model}
\usage{
heatmapLCN(mod, grp, minGrpSize = NULL, prob_cols = c("black", "grey",
  "blue"), greater_col = "red", plotChannelNumber = T, xlab = " ",
  ylab = " ", sortColumns = T, ...)
}
\arguments{
\item{mod}{LCN or BKN model}

\item{grp}{Vector of group categories for each node}

\item{minGrpSize}{Minimum size of group in both. Smaller groups put in "other"}

\item{prob_cols}{Colors for color gradient of probability range}

\item{greater_col}{Color for color gradient beyond 1}

\item{xlab}{X-axis label}

\item{ylab}{Y-asix label}

\item{...}{Additional arguments passed to ComplexHeatmap::Heatmap}

\item{reorderChannels}{Should Channels be reorder by dependency on grp?}
}
\description{
Build heatmap from model
}
