% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModelWrappers.R
\name{makeLCN}
\alias{makeLCN}
\title{Make Latent Channel Network Model}
\usage{
makeLCN(edgeList, nDims = 5)
}
\arguments{
\item{edgeList}{A nx2 matrix of edges}

\item{nDims}{Number of Latent Channels}
}
\description{
Make Latent Channel Network Model
}
