% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_net}
\alias{plot_net}
\title{Build heatmap from model}
\usage{
plot_net(
  mod,
  grp = NULL,
  metanames = NULL,
  minGrpSize = NULL,
  row_subset = NULL,
  col_subset = NULL,
  name = "",
  plotratio = 2,
  ...
)
}
\arguments{
\item{mod}{LatClass object}

\item{grp}{Vector of group categories for each node}

\item{metanames}{Names of metavariables to plot}

\item{minGrpSize}{Minimum size of group in both. Smaller groups put in "other"}

\item{row_subset}{Subset of nodes to plot}

\item{col_subset}{Subset of channels to plot}

\item{name}{Legend names for plot}

\item{plotratio}{If node parameters + meta parameters plotted, ratio between plots}

\item{...}{Additional arguments passed to ComplexHeatmap::Heatmap}
}
\description{
Build heatmap from model
}
