% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatClass.R
\name{predict.LatClass}
\alias{predict.LatClass}
\title{Predictions from LatClass objects}
\usage{
\method{predict}{LatClass}(mod, i, j, type = "pairs")
}
\arguments{
\item{mod}{LatClass model}

\item{i}{node index}

\item{j}{Either an node index or metadata colname name}

\item{type}{Should node pairs ('pairs') or cross ('cross') of all combinations be predicted}
}
\description{
Predict edge probabilities and categorical metadata
}
\examples{
data(email_data)

# Building model and fitting
mod = makeLatentModel(email_data$edgeList, 
                      nChans = 10, 
                      metadata = email_data$meta)
mod$fit(fast_em = TRUE)

# Predicting edge pairs
predict(mod, i = 1:3, j = 4:2)

# Predicting all combinations of i and j
predict(mod, i = 1:3, j = 1:3, type = "cross")

# Predicting metadata 
# Subsetting for brevity
predict(mod, i = 1:3, "dpt")[,1:5]
}
