% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModelWrappers.R
\name{prepEdgeList}
\alias{prepEdgeList}
\title{Prep Edgelist}
\usage{
prepEdgeList(edgeList)
}
\arguments{
\item{edgeList}{An nx2 matrix of undirected edge pairs}
}
\description{
Prepares edgelist for \code{makeLCN}
}
